Implement API operations

+++ src/Neo4jManager/Neo4jManager/V2/Neo4jV2JavaInstanceProvider.cs
@@ -16,8 +16,8 @@ namespace Neo4jManager.V2
         private Process process;
 
 
-        public Neo4jV2JavaInstanceProvider(string javaPath, string neo4jHomeFolder, IFileCopy fileCopy, Neo4jEndpoints endpoints)
-            :base(neo4jHomeFolder, fileCopy, endpoints)
+        public Neo4jV2JavaInstanceProvider(string javaPath, string neo4jHomeFolder, IFileCopy fileCopy, Neo4jVersion neo4jVersion, Neo4jEndpoints endpoints)
+            :base(neo4jHomeFolder, fileCopy, neo4jVersion, endpoints)
         {
             this.javaPath = javaPath;
         }
@@ -41,8 +41,6 @@ namespace Neo4jManager.V2
 
         public override async Task Stop(CancellationToken token)
         {
-            if (process == null || process.HasExited) return;
-
             await Task.Run(() =>
             {
                 Stop();
@@ -51,6 +49,8 @@ namespace Neo4jManager.V2
 
         private void Stop()
         {
+            if (process == null || process.HasExited) return;
+
             process.Kill();
             process.WaitForExit(defaultWaitForKill);
         }
